+++ deepreg/train.py
@@ -144,7 +144,7 @@ def train(
     tensorboard_callback = tf.keras.callbacks.TensorBoard(
         log_dir=log_dir,
         histogram_freq=config["train"]["save_period"],
-        update_freq=100,
+        update_freq=config["train"].get("update_freq", "epoch"),
     )
     ckpt_callback, initial_epoch = build_checkpoint_callback(
         model=model,